requirements.txt
numpy==1.20.2
streamlit==0.82.0
antropy==0.1.4
pandas==1.2.3
scipy==1.6.2
Pillow==8.2.0
scikit_learn==0.24.2
